+++ docs/source/release-notes.rst
@@ -9,6 +9,15 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 v1.9 Notes
 ==========
 
+v1.9.2 (2024-03-25)
+-------------------
+- Improve PSS/E parser for the `wmod` field in the static generator
+  section.
+- Consider line status when parsing PSS/E file.
+- Added functions in `Line` for building network admittance matrix, `Bdc`
+  matrix for DC power flow, and `Bp` and `Bpp` matrices for fast decoupled
+  power flow. See ``build_y``, ``build_b`` and ``build_Bdc``.
+
 v1.9.1 (2024-02-04)
 -------------------
 This is a hotfix to pass ``pip check`` for KVXOPT version.